Adding search feature, deleteall should delete only filtered links

+++ js/popup.js
@@ -20,8 +20,9 @@ chrome.tabs.onActivated.addListener(function(tabId, changeInfo, tab) {
 });
 
 //store
-let links = [];
-let db = new Dexie("one_tab");
+let allLinks = [];
+let filteredLinks = [];
+let db = new Dexie("tab_wiz");
 db.version(1).stores({
   links: "url,title,favIconUrl"
 });
@@ -69,7 +70,7 @@ function generateListHTML(list) {
   return listHtml;
 }
 
-function setListView() {
+function setListView(links) {
   let list = document.getElementById("list");
   list.innerHTML = generateListHTML(links);
   let deleteBtns = document.querySelectorAll(".delete-icon");
@@ -87,7 +88,9 @@ function setListView() {
     );
   });
 
-  let appControlBtns = document.querySelectorAll(".app-controls button");
+  let appControlBtns = document.querySelectorAll(
+    ".app-controls .can-be-disabled"
+  );
   if (links.length) {
     appControlBtns.forEach(btn => {
       btn.classList.remove("disabled");
@@ -106,11 +109,11 @@ function closeTab(tabID) {
 }
 
 function linkExists(link) {
-  return links.find(item => item.url == link.url) ? true : false;
+  return allLinks.find(item => item.url == link.url) ? true : false;
 }
 async function init() {
-  links = await linkStore.toArray();
-  setListView();
+  allLinks = await linkStore.toArray();
+  setListView(allLinks);
 }
 init();
 async function addLink(link) {
@@ -119,8 +122,8 @@ async function addLink(link) {
   if (!linkExists(link)) {
     try {
       await save(link);
-      links.push(link);
-      setListView();
+      allLinks.push(link);
+      setListView(allLinks);
     } catch (err) {
       console.log("Could not save to db");
     }
@@ -133,26 +136,33 @@ async function addLinks(tabLinks) {
     closeTab(tab.id);
   });
   await init();
-  let linkUrls = links.map(link => link.url);
+  let linkUrls = allLinks.map(link => link.url);
   let newLinks = tabLinks.filter(link => !linkUrls.includes(link.url));
   let newUrls = tabLinks.map(link => link.url); //temp array containing just urls
   newLinks = newLinks.filter((link, index) => {
     return newUrls.indexOf(link.url) == index; // if index is not same that means it is not unique
   });
   await saveAll(newLinks);
-  links = links.concat(newLinks);
-  setListView();
+  allLinks = allLinks.concat(newLinks);
+  setListView(allLinks);
 }
 async function deleteLink() {
   try {
     await deleteFromDB(this.url);
-    links = links.filter(link => link.url != this.url);
-    setListView();
+    allLinks = allLinks.filter(link => link.url != this.url);
+    setListView(allLinks);
   } catch (err) {
     console.log("error while deleting");
   }
 }
+let searchBox = document.getElementById("searchInput");
+
+//checks if there is a filtered links list, if so returns it else returns allLinks list
+function isFilteredList() {
+  return filteredLinks.length && filteredLinks.length != allLinks.length;
+}
 async function deleteAll() {
+  let links = isFilteredList() ? filteredLinks : allLinks;
   if (links.length) {
     //only if there are values in the links array
     let deleteConfirm;
@@ -162,24 +172,24 @@ async function deleteAll() {
     if (deleteConfirm || !this.showPrompt) {
       // showPrompt if not set will delete without confirm box
       try {
-        await deleteAllFromDB();
-        links = [];
-        setListView();
+        if (isFilteredList()) {
+          links.forEach(async link => {
+            let url = link.url;
+            await deleteLink.bind({ url })();
+          });
+          searchBox.value = "";
+        } else {
+          await deleteAllFromDB();
+        }
+        //reset
+        filteredLinks= [];
+        links= [];
+        
+        setListView(allLinks);
       } catch (err) {
         console.log(err);
       }
     }
-  }
-}
-async function restoreAll() {
-  if (links.length) {
-    links.forEach(link => {
-      chrome.tabs.create({
-        url: link.url
-      });
-    });
-    await deleteAll();
-    window.close();
   }
 }
 
@@ -191,14 +201,68 @@ deleteAllBtn.addEventListener(
   })
 );
 
+async function restoreAll() {
+  if (allLinks.length) {
+    allLinks.forEach(link => {
+      chrome.tabs.create({
+        url: link.url
+      });
+    });
+    await deleteAll();
+    window.close();
+  }
+}
+
 let restoreAllBtn = document.getElementById("restoreAll");
 restoreAllBtn.addEventListener("click", restoreAll);
 
 let exportBtn = document.getElementById("export");
 exportBtn.addEventListener("click", function() {
-  var file = new Blob([JSON.stringify(links)], { type: "text/json" });
+  var file = new Blob([JSON.stringify(allLinks)], { type: "text/json" });
   var a = document.createElement("a");
   a.href = URL.createObjectURL(file);
   a.download = "tabs.json"; //file name
   a.click();
+});
+
+let importBtn = document.getElementById("import");
+importBtn.addEventListener("click", function() {
+  let fileInput = document.createElement("input");
+  fileInput.type = "file";
+  fileInput.click();
+  fileInput.addEventListener("change", function(data) {
+    console.log(data);
+    let file = fileInput.files[0];
+    console.log(file);
+    var reader = new FileReader(file);
+    reader.readAsText(file);
+    reader.onload = async function(e) {
+      // browser completed reading file - display it
+      let importLinks = e.target.result;
+      try {
+        importLinks = JSON.parse(importLinks);
+        console.log(importLinks);
+        //Todo: validation
+        await saveAll(importLinks);
+        allLinks = allLinks.concat(importLinks);
+        setListView(allLinks);
+      } catch (err) {
+        console.log(err);
+
+        alert("File type is not supported");
+      }
+    };
+  });
+});
+
+
+searchBox.addEventListener("keyup", function(event) {
+  console.log("key", event.target.value);
+  let searchText = event.target.value.toLocaleLowerCase();
+  searchText
+    ? (filteredLinks = allLinks.filter(link =>
+        link.title.toLocaleLowerCase().includes(searchText)
+      ))
+    : (filteredLinks = allLinks);
+  setListView(filteredLinks);
 });
